changed div to footer and extracted for reuse
